requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==0.8.11
google-generativeai==0.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 